// src/types.ts
export type NavModule =
  | 'dashboard'
  | 'tasks'
  | 'routine'
  | 'goals'
  | 'focus'
  | 'notes'
  | 'journal'
  | 'calculator'
  | 'finance'
  | 'health'
  | 'work'
  | 'people'
  | 'spiritual'
  | 'things'
  | 'settings';